Extract Witness types from shelleyma `Tx` module into new module `TxWits`
